Refactor DocBuilder.Generate() spin impl to own file
This commit splits DocBuilder.Generate() into more scoped functions
to reduce its complexity. The implementation is now more mockable and
the split functions should prove to be more testable.
