+++ pkg/spdx/builder.go
@@ -22,13 +22,7 @@ import (
 	"net/url"
 	"os"
 	"path/filepath"
-	"strings"
 
-	"github.com/google/uuid"
-	"github.com/sirupsen/logrus"
-	"gopkg.in/yaml.v2"
-
-	"sigs.k8s.io/bom/pkg/license"
 	"sigs.k8s.io/release-utils/util"
 )
 
@@ -81,24 +75,56 @@ func NewDocBuilder(options ...NewDocBuilderOption) *DocBuilder {
 	return db
 }
 
-// DocBuilder is a tool to write spdx manifests
+// DocBuilder is a tool to write SPDX SBOMs. It is configurable by
+// defining values in its DocBuilderOptions. Options to customize the
+// generated document are passed to the Generate() method in DocGenerateOptions
+// struct.
 type DocBuilder struct {
 	options *DocBuilderOptions
 	impl    DocBuilderImplementation
 }
 
-// Generate creates anew SPDX document describing the artifacts specified in the options
+// Generate creates a new SPDX SBOM. The resulting document will describe the all
+// artifacts specified in the DocGenerateOptions struct passed.
 func (db *DocBuilder) Generate(genopts *DocGenerateOptions) (*Document, error) {
-	if genopts.ConfigFile != "" {
-		if err := db.impl.ReadYamlConfiguration(genopts.ConfigFile, genopts); err != nil {
-			return nil, fmt.Errorf("parsing configuration file: %w", err)
-		}
+	if err := db.impl.ReadYamlConfiguration(genopts.ConfigFile, genopts); err != nil {
+		return nil, fmt.Errorf("parsing configuration file: %w", err)
 	}
-	// Create the SPDX document
-	doc, err := db.impl.GenerateDoc(db.options, genopts)
+
+	if err := db.impl.ValidateOptions(genopts); err != nil {
+		return nil, fmt.Errorf("checking build options: %w", err)
+	}
+
+	spdx, err := db.impl.CreateSPDXClient(genopts, db.options)
 	if err != nil {
-		return nil, fmt.Errorf("creating SPDX document: %w", err)
+		return nil, fmt.Errorf("generating spdx client")
 	}
+
+	doc, err := db.impl.CreateDocument(genopts, spdx)
+	if err != nil {
+		return nil, fmt.Errorf("creating spdx document: %w", err)
+	}
+
+	if err := db.impl.ScanDirectories(genopts, spdx, doc); err != nil {
+		return nil, fmt.Errorf("scanning directories: %w", err)
+	}
+
+	if err := db.impl.ScanImages(genopts, spdx, doc); err != nil {
+		return nil, fmt.Errorf("scanning images: %w", err)
+	}
+
+	if err := db.impl.ScanImageArchives(genopts, spdx, doc); err != nil {
+		return nil, fmt.Errorf("scanning image archives: %w", err)
+	}
+
+	if err := db.impl.ScanArchives(genopts, spdx, doc); err != nil {
+		return nil, fmt.Errorf("scanning archives: %w", err)
+	}
+
+	if err := db.impl.ScanFiles(genopts, spdx, doc); err != nil {
+		return nil, fmt.Errorf("scanning files: %w", err)
+	}
+
 	return doc, nil
 }
 
@@ -156,158 +182,6 @@ var defaultDocBuilderOpts = DocBuilderOptions{
 	WorkDir: filepath.Join(os.TempDir(), "spdx-docbuilder"),
 }
 
-type DocBuilderImplementation interface {
-	GenerateDoc(*DocBuilderOptions, *DocGenerateOptions) (*Document, error)
-	WriteDoc(*Document, string) error
-	ReadYamlConfiguration(string, *DocGenerateOptions) error
-}
-
-// defaultDocBuilderImpl is the default implementation for the
-// SPDX document builder
-type defaultDocBuilderImpl struct {
-	format Format
-}
-
-// Generate generates a document
-func (builder *defaultDocBuilderImpl) GenerateDoc(
-	opts *DocBuilderOptions, genopts *DocGenerateOptions,
-) (doc *Document, err error) {
-	if err := genopts.Validate(); err != nil {
-		return nil, fmt.Errorf("checking build options: %w", err)
-	}
-
-	spdx := NewSPDX()
-	if len(genopts.IgnorePatterns) > 0 {
-		spdx.Options().IgnorePatterns = genopts.IgnorePatterns
-	}
-	spdx.Options().AnalyzeLayers = genopts.AnalyseLayers
-	spdx.Options().ProcessGoModules = genopts.ProcessGoModules
-	spdx.Options().ScanImages = genopts.ScanImages
-	spdx.Options().LicenseListVersion = genopts.LicenseListVersion
-
-	if !util.Exists(opts.WorkDir) {
-		if err := os.MkdirAll(opts.WorkDir, os.FileMode(0o755)); err != nil {
-			return nil, fmt.Errorf("creating builder worskpace dir: %w", err)
-		}
-	}
-
-	// Create the new document
-	doc = NewDocument()
-	doc.Name = genopts.Name
-	doc.LicenseListVersion = strings.TrimPrefix(license.DefaultCatalogOpts.Version, "v")
-	if genopts.LicenseListVersion != "" {
-		doc.LicenseListVersion = strings.TrimPrefix(genopts.LicenseListVersion, "v")
-	}
-
-	// If we do not have a namespace, we generate one
-	// under the public SPDX URL defined in the spec.
-	// (ref https://spdx.github.io/spdx-spec/document-creation-information/#65-spdx-document-namespace-field)
-	if genopts.Namespace == "" {
-		doc.Namespace = "https://spdx.org/spdxdocs/k8s-releng-bom-" + uuid.NewString()
-	} else {
-		doc.Namespace = genopts.Namespace
-	}
-	doc.Creator.Person = genopts.CreatorPerson
-	doc.ExternalDocRefs = genopts.ExternalDocumentRef
-
-	for _, dirPattern := range genopts.Directories {
-		matches, err := filepath.Glob(dirPattern)
-		if err != nil {
-			return nil, err
-		}
-		for _, dirMatch := range matches {
-			isFile, err := pathIsOfFile(dirMatch)
-			if err != nil {
-				return nil, fmt.Errorf("stat dir: %w", err)
-			}
-			if isFile {
-				logrus.Debugf("Skipping %s because it's a file", dirMatch)
-				continue
-			}
-			logrus.Infof("Processing directory %s", dirMatch)
-			pkg, err := spdx.PackageFromDirectory(dirMatch)
-			if err != nil {
-				return nil, fmt.Errorf("generating package from directory: %w", err)
-			}
-			doc.ensureUniqueElementID(pkg)
-			if err := doc.AddPackage(pkg); err != nil {
-				return nil, fmt.Errorf("adding directory package to document: %w", err)
-			}
-		}
-	}
-
-	// Process all image references from registries
-	for _, i := range genopts.Images {
-		logrus.Infof("Processing image reference: %s", i)
-		p, err := spdx.ImageRefToPackage(i)
-		if err != nil {
-			return nil, fmt.Errorf("generating SPDX package from image ref %s: %w", i, err)
-		}
-		doc.ensureUniqueElementID(p)
-		doc.ensureUniquePeerIDs(p.GetRelationships())
-		if err := doc.AddPackage(p); err != nil {
-			return nil, fmt.Errorf("adding package to document: %w", err)
-		}
-	}
-
-	// Process OCI image archives
-	for _, tb := range genopts.Tarballs {
-		logrus.Infof("Processing image archive %s", tb)
-		p, err := spdx.PackageFromImageTarball(tb)
-		if err != nil {
-			return nil, fmt.Errorf("generating tarball package: %w", err)
-		}
-		doc.ensureUniqueElementID(p)
-		doc.ensureUniquePeerIDs(p.GetRelationships())
-		if err := doc.AddPackage(p); err != nil {
-			return nil, fmt.Errorf("adding package to document: %w", err)
-		}
-	}
-
-	// Add archive files as packages
-	for _, tf := range genopts.Archives {
-		logrus.Infof("Adding archive file as package: %s", tf)
-		p, err := spdx.PackageFromArchive(tf)
-		if err != nil {
-			return nil, fmt.Errorf("creating spdx package from archive: %w", err)
-		}
-		doc.ensureUniqueElementID(p)
-		doc.ensureUniquePeerIDs(p.GetRelationships())
-		if err := doc.AddPackage(p); err != nil {
-			return nil, fmt.Errorf("adding package to document: %w", err)
-		}
-	}
-
-	// Process single files, not part of a package
-	for _, filePattern := range genopts.Files {
-		matches, err := filepath.Glob(filePattern)
-		if err != nil {
-			return nil, err
-		}
-		if len(matches) == 0 {
-			logrus.Warnf("%s pattern didn't match any file", filePattern)
-		}
-		for _, filePath := range matches {
-			isFile, err := pathIsOfFile(filePath)
-			if err != nil {
-				return nil, fmt.Errorf("stat file: %w", err)
-			}
-			if !isFile {
-				continue
-			}
-			f, err := spdx.FileFromPath(filePath)
-			if err != nil {
-				return nil, fmt.Errorf("adding file: %w", err)
-			}
-			doc.ensureUniqueElementID(f)
-			if err := doc.AddFile(f); err != nil {
-				return nil, fmt.Errorf("adding file to document: %w", err)
-			}
-		}
-	}
-	return doc, nil
-}
-
 // TODO: Move this to https://github.com/kubernetes-sigs/release-utils/blob/main/util/common.go#L485
 func pathIsOfFile(path string) (bool, error) {
 	fInfo, err := os.Stat(path)
@@ -315,74 +189,4 @@ func pathIsOfFile(path string) (bool, error) {
 		return false, err
 	}
 	return !fInfo.IsDir(), nil
-}
-
-// WriteDoc renders the document to a file
-func (builder *defaultDocBuilderImpl) WriteDoc(doc *Document, path string) error {
-	markup, err := doc.Render()
-	if err != nil {
-		return fmt.Errorf("generating document markup: %w", err)
-	}
-	logrus.Infof("writing document to %s", path)
-
-	if err := os.WriteFile(path, []byte(markup), os.FileMode(0o644)); err != nil {
-		return fmt.Errorf(
-			"writing document markup to file: %w",
-			err,
-		)
-	}
-	return nil
-}
-
-// ReadYamlConfiguration reads a yaml configuration and
-// set the values in an options struct
-func (builder *defaultDocBuilderImpl) ReadYamlConfiguration(
-	path string, opts *DocGenerateOptions,
-) (err error) {
-	yamldata, err := os.ReadFile(path)
-	if err != nil {
-		return fmt.Errorf("reading yaml SBOM configuration: %w", err)
-	}
-
-	conf := &YamlBOMConfiguration{}
-	if err := yaml.Unmarshal(yamldata, conf); err != nil {
-		return fmt.Errorf("unmarshalling SBOM configuration YAML: %w", err)
-	}
-
-	if conf.Name != "" {
-		opts.Name = conf.Name
-	}
-
-	if conf.Namespace != "" {
-		opts.Namespace = conf.Namespace
-	}
-
-	if conf.Creator.Person != "" {
-		opts.CreatorPerson = conf.Creator.Person
-	}
-
-	if conf.License != "" {
-		opts.License = conf.License
-	}
-
-	opts.ExternalDocumentRef = conf.ExternalDocRefs
-
-	// Add all the artifacts
-	for _, artifact := range conf.Artifacts {
-		logrus.Infof("Configuration has artifact of type %s: %s", artifact.Type, artifact.Source)
-		switch artifact.Type {
-		case "directory":
-			opts.Directories = append(opts.Directories, artifact.Source)
-		case "image":
-			opts.Images = append(opts.Images, artifact.Source)
-		case "docker-archive":
-			opts.Tarballs = append(opts.Tarballs, artifact.Source)
-		case "file":
-			opts.Files = append(opts.Files, artifact.Source)
-		case "archive":
-			opts.Archives = append(opts.Archives, artifact.Source)
-		}
-	}
-
-	return nil
 }
